refactor: consolidate manual commands into single JSON file
Merge client/manualCommands.js and server/game/manualCommands.js into a
shared manualCommands.json at the repo root. Update HowToPlay.jsx and
chatcommands.js to import from the new location.
