+++ .gitignore
@@ -8,6 +8,7 @@ coverage
 *.log
 *.json
 !client/locales/*.json
+!manualCommands.json
 server/events.txt
 .DS_Store
 config/local*
